refactor(fieldmap): extract FieldMapTooltip into a separate component
