+++ js/source/include/fieldmap/components/FieldMapLegend.tsx
