+++ js/source/include/fieldmap/hooks/useZoomPan.ts
